+++ components/analytics/shake2report/api/src/main/java/com/flipperdevices/shake2report/api/Shake2ReportDecomposeComponent.kt
@@ -1,0 +1,12 @@
+package com.flipperdevices.shake2report.api
+
+import com.flipperdevices.ui.decompose.DecomposeComponent
+import com.flipperdevices.ui.decompose.DecomposeOnBackParameter
+
+interface Shake2ReportDecomposeComponent : DecomposeComponent {
+    fun interface Factory {
+        operator fun invoke(
+            onBack: DecomposeOnBackParameter
+        ): Shake2ReportDecomposeComponent
+    }
+}