+++ components/core/ui/decompose/src/main/java/com/flipperdevices/ui/decompose/DecomposeOnBackParameter.kt
@@ -1,0 +1,5 @@
+package com.flipperdevices.ui.decompose
+
+fun interface DecomposeOnBackParameter {
+    operator fun invoke()
+}