+++ components/debug/api/src/main/java/com/flipperdevices/debug/api/StressTestDecomposeComponent.kt
@@ -1,0 +1,5 @@
+package com.flipperdevices.debug.api
+
+import com.flipperdevices.ui.decompose.DecomposeComponent
+
+interface StressTestDecomposeComponent : DecomposeComponent