+++ components/filemanager/api/src/main/java/com/flipperdevices/filemanager/api/navigation/FileManagerDecomposeComponent.kt
@@ -1,0 +1,12 @@
+package com.flipperdevices.filemanager.api.navigation
+
+import com.arkivanov.decompose.ComponentContext
+import com.flipperdevices.ui.decompose.DecomposeComponent
+
+interface FileManagerDecomposeComponent : DecomposeComponent {
+    fun interface Factory {
+        operator fun invoke(
+            componentContext: ComponentContext,
+        ): FileManagerDecomposeComponent
+    }
+}