+++ components/settings/api/src/main/java/com/flipperdevices/settings/api/SettingsFeatureEntry.kt
@@ -6,6 +6,4 @@ import com.flipperdevices.core.ui.navigation.FeatureScreenRootRoute
 interface SettingsFeatureEntry : AggregateFeatureEntry {
     override val ROUTE: FeatureScreenRootRoute
         get() = FeatureScreenRootRoute.OPTIONS
-
-    fun remoteControlStub(): String
 }