+++ components/settings/impl/src/main/java/com/flipperdevices/settings/impl/viewmodels/ExperimentalViewModel.kt
@@ -1,23 +1,22 @@
 package com.flipperdevices.settings.impl.viewmodels
 
-import androidx.navigation.NavController
+import com.arkivanov.decompose.router.stack.StackNavigation
+import com.arkivanov.decompose.router.stack.push
 import com.flipperdevices.core.di.provideDelegate
 import com.flipperdevices.core.ui.lifecycle.LifecycleViewModel
-import com.flipperdevices.filemanager.api.navigation.FileManagerEntry
 import com.flipperdevices.metric.api.MetricApi
 import com.flipperdevices.metric.api.events.SimpleEvent
-import tangle.viewmodel.VMInject
+import com.flipperdevices.settings.impl.model.SettingsNavigationConfig
+import javax.inject.Inject
 import javax.inject.Provider
 
-class ExperimentalViewModel @VMInject constructor(
-    fileManagerEntryProvider: Provider<FileManagerEntry>,
+class ExperimentalViewModel @Inject constructor(
     metricApiProvider: Provider<MetricApi>
 ) : LifecycleViewModel() {
-    private val fileManagerEntry by fileManagerEntryProvider
     private val metricApi by metricApiProvider
 
-    fun onOpenFileManager(navController: NavController) {
+    fun onOpenFileManager(navigation: StackNavigation<SettingsNavigationConfig>) {
         metricApi.reportSimpleEvent(SimpleEvent.EXPERIMENTAL_OPEN_FM)
-        navController.navigate(fileManagerEntry.fileManagerDestination())
+        navigation.push(SettingsNavigationConfig.FileManager)
     }
 }